+++ src/Ops.Data/Ops/CoverIssueHeaderRepository.cs
@@ -7,7 +7,6 @@ using Ocuda.Ops.Data.Extensions;
 using Ocuda.Ops.Models.Entities;
 using Ocuda.Ops.Service.Filters;
 using Ocuda.Ops.Service.Interfaces.Ops.Repositories;
-using Ocuda.Ops.Service.Models;
 using Ocuda.Utility.Models;
 
 namespace Ocuda.Ops.Data.Ops