+++ src/Ops.Data/OpsContext.cs
@@ -60,6 +60,7 @@ namespace Ocuda.Ops.Data
         public DbSet<FileType> FileTypes { get; set; }
         public DbSet<Link> Links { get; set; }
         public DbSet<LinkLibrary> LinkLibraries { get; set; }
+        public DbSet<PermissionGroup> PermissionGroups { get; set; }
         public DbSet<PostCategory> PostCategories { get; set; }
         public DbSet<Post> Posts { get; set; }
         public DbSet<Category> Categories { get; set; }