+++ src/Ops.Service/CoverIssueService.cs
@@ -8,7 +8,7 @@ using Ocuda.Ops.Service.Abstract;
 using Ocuda.Ops.Service.Filters;
 using Ocuda.Ops.Service.Interfaces.Ops.Repositories;
 using Ocuda.Ops.Service.Interfaces.Ops.Services;
-using Ocuda.Ops.Service.Models;
+using Ocuda.Utility.Models;
 
 namespace Ocuda.Ops.Service
 {