+++ src/Ops.Service/Interfaces/Ops/Repositories/IExternalResourceRepository.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using Ocuda.Ops.Models.Entities;
 using Ocuda.Ops.Service.Filters;
-using Ocuda.Ops.Service.Models;
 using Ocuda.Utility.Models;
 
 namespace Ocuda.Ops.Service.Interfaces.Ops.Repositories