+++ src/Ops.Service/Interfaces/Ops/Repositories/IUserMetadataTypeRepository.cs
@@ -2,14 +2,15 @@
 using System.Threading.Tasks;
 using Ocuda.Ops.Models.Entities;
 using Ocuda.Ops.Service.Filters;
-using Ocuda.Ops.Service.Models;
+using Ocuda.Utility.Models;
 
 namespace Ocuda.Ops.Service.Interfaces.Ops.Repositories
 {
     public interface IUserMetadataTypeRepository : IOpsRepository<UserMetadataType, int>
     {
         Task<ICollection<UserMetadataType>> GetAllAsync();
-        Task<DataWithCount<ICollection<UserMetadataType>>> GetPaginatedListAsync(BaseFilter filter);
+        Task<DataWithCount<ICollection<UserMetadataType>>>
+            GetPaginatedListAsync(BaseFilter filter);
         Task<bool> IsDuplicateAsync(UserMetadataType metadataType);
     }
 }