+++ src/Promenade.Models/Entities/PodcastDirectory.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Text;
+
+namespace Ocuda.Promenade.Models.Entities
+{
+    public class PodcastDirectory
+    {
+        [Key]
+        [Required]
+        public int Id { get; set; }
+
+        [Required]
+        [MaxLength(255)]
+        public string Name { get; set; }
+    }
+}