+++ src/Promenade.Web/Startup.cs
@@ -262,6 +262,10 @@ namespace Ocuda.Promenade.Web
                 Data.Promenade.NavigationTextRepository>();
             services.AddScoped<Service.Interfaces.Repositories.IPageRepository,
                 Data.Promenade.PageRepository>();
+            services.AddScoped<Service.Interfaces.Repositories.IPodcastItemRepository,
+                Data.Promenade.PodcastItemRepository>();
+            services.AddScoped<Service.Interfaces.Repositories.IPodcastRepository,
+                Data.Promenade.PodcastRepository>();
             services.AddScoped<Service.Interfaces.Repositories.IScheduleRequestRepository,
                 Data.Promenade.ScheduleRequestRepository>();
             services.AddScoped<Service.Interfaces.Repositories.IScheduleRequestSubjectRepository,
@@ -296,6 +300,7 @@ namespace Ocuda.Promenade.Web
             services.AddScoped<NavigationService>();
             services.AddScoped<PageService>();
             services.AddScoped<RedirectService>();
+            services.AddScoped<PodcastService>();
             services.AddScoped<ScheduleService>();
             services.AddScoped<SegmentService>();
             services.AddScoped<SiteAlertService>();